src/requirements.txt
boto3==1.9.0
botocore==1.12.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
dj-database-url==0.5.0
Django==2.0.5
django-cors-headers==2.4.0
django-storages==1.7.1
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
docutils==0.14
gunicorn==19.9.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
stripe==2.8.0
uritemplate==3.0.0
urllib3==1.22
